requirements.txt
transformers
torch
requests
python-telegram-bot
feedparser
python-dotenv
